# Conflicts:
#   frontend/src/pages/DesertionPage.js

Former-commit-id: a54e139ecf2cf1034efc823ace45bc0302d9de5e
Former-commit-id: c75ec2156674730f37fe63910fd55eb2855131ef

+++ backend/src/main/java/com/moi/anitime/model/repo/BookmarkRepo.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 public interface BookmarkRepo extends JpaRepository<Bookmark, Integer> {
 
     @Query(value = "SELECT b from Bookmark b WHERE b.animal.desertionNo = :desertionNo AND b.generalMember.memberNo = :generalNo")
-    public Optional<Bookmark> findBookmark(@Param("desertionNo") int desertionNo, @Param("generalNo") int generalNo);
+    public Optional<Bookmark> findBookmark(@Param("desertionNo") long desertionNo, @Param("generalNo") int generalNo);
 
 
 }
